+++ dbmodel/utils/make.py
@@ -7,6 +7,19 @@ import re
 import mysql.connector as mariadb
 from dbmodel.utils.inflector import Inflector, Portugues
 
+def reference_name(table, list, name, fk_name):
+    __table_referenced = name
+    if not __table_referenced in list:
+        list.append(__table_referenced)
+        return __table_referenced
+    else:
+        __table_referenced = fk_name
+        count = 0
+        while __table_referenced in list:
+            count = count + 1
+            __table_referenced = "{}{}".format(__table_referenced, count)
+        list.append(__table_referenced)
+        return __table_referenced
 
 def Make(dir, db_user, db_password, db_host, db_port, db_database, db_ssl=False, db_ssl_ca=None, db_ssl_cert=None, db_ssl_key=None, date_format="%d/%m/%Y %H:%M:%S"):
 
@@ -70,14 +83,12 @@ def Make(dir, db_user, db_password, db_host, db_port, db_database, db_ssl=False,
         table_relationships_n_to_n = []
 
         # CHECK AND MAKE LIST FOR MANY TO MANY RELATION
-        if table["TABLE_NAME"] == "usuarios":
-            for to_many_relation in table_relationships_one_to_n:
-                table_columns_to_many = [column for column in columns if column["TABLE_NAME"] == to_many_relation["table_name"]]
-                # CHECK IF HAS 2 FIELDS AND ALL FIELDS ARE PRIMARY AND NOT NULL
-                if len(table_columns_to_many) == 2 and all([col["COLUMN_KEY"] == 'PRI' and col["IS_NULLABLE"]=="NO" for col in table_columns_to_many]):
-                    table_relationships_to_many = [
-                        relationship for relationship in relationships if relationship["table_name"] == to_many_relation["table_name"] and relationship["referenced_table_name"] != table["TABLE_NAME"]][0]
-                    table_relationships_n_to_n.append({"name": table_relationships_to_many["referenced_table_name"], "table": table_relationships_to_many["referenced_table_name"], "intermediate": to_many_relation["table_name"], "key": to_many_relation["column_name"], "reference": to_many_relation["referenced_column_name"], "inter_key" : table_relationships_to_many["column_name"], "end_key": table_relationships_to_many["referenced_column_name"]})
+        for to_many_relation in table_relationships_one_to_n:
+            table_columns_to_many = [column for column in columns if column["TABLE_NAME"] == to_many_relation["table_name"]]
+            # CHECK IF HAS 2 FIELDS AND ALL FIELDS ARE PRIMARY AND NOT NULL
+            if len(table_columns_to_many) == 2 and all([col["COLUMN_KEY"] == 'PRI' and col["IS_NULLABLE"]=="NO" for col in table_columns_to_many]):
+                table_relationships_to_many = [relationship for relationship in relationships if relationship["table_name"] == to_many_relation["table_name"] and relationship["referenced_table_name"] != table["TABLE_NAME"]][0]
+                table_relationships_n_to_n.append({"name": table_relationships_to_many["referenced_table_name"], "table": table_relationships_to_many["referenced_table_name"], "intermediate": to_many_relation["table_name"], "key": to_many_relation["column_name"], "reference": to_many_relation["referenced_column_name"], "inter_key" : table_relationships_to_many["column_name"], "end_key": table_relationships_to_many["referenced_column_name"], "constraint_name": table_relationships_to_many["constraint_name"]})
 
 
         table_columns = [
@@ -86,8 +97,10 @@ def Make(dir, db_user, db_password, db_host, db_port, db_database, db_ssl=False,
         path_entitie = "{}/{}.py".format(model_path, classname.lower())
         models.append("from model.{} import {}".format(
             classname.lower(), classname))
+
         print("\t\t\u2714 Creating Model: {} : {}".format(
             table["TABLE_NAME"], classname))
+
         with open(path_entitie, "w") as entitie:
             entitie.write("# -*- coding: utf-8 -*-")
             entitie.write("\nfrom dbmodel.entity import *")
@@ -141,40 +154,47 @@ def Make(dir, db_user, db_password, db_host, db_port, db_database, db_ssl=False,
 
             for table_relationship in table_relationships_one_to_one:
                 if table_relationship["referenced_table_name"] != table["TABLE_NAME"]:
-                    referenced_table.append(table_relationship["referenced_table_name"])
+
+                    __table_referenced = reference_name(table["TABLE_NAME"], referenced_table, table_relationship["referenced_table_name"], table_relationship["constraint_name"])
+
                     entitie.write("\n\n\t@Object(name=\"{}\", key=\"{}\", reference=\"{}\", table=\"{}\")".format(_inflector.classify(
                         table_relationship["referenced_table_name"]), table_relationship["referenced_column_name"], table_relationship["column_name"], table_relationship["referenced_table_name"]))
-                    entitie.write("\n\tdef {}(self): pass".format(
-                        table_relationship["referenced_table_name"]))
+
+                    entitie.write("\n\tdef {}(self): pass".format(__table_referenced))
 
             if len(table_relationships_one_to_n) > 0:
                 entitie.write("\n\n\t# One-to-many")
 
             for table_relationship in table_relationships_one_to_n:
 
+                __table_referenced = reference_name(table["TABLE_NAME"], referenced_table, table_relationship["table_name"], table_relationship["constraint_name"])
+
                 entitie.write("\n\n\t@ObjectList(name=\"{}\", key=\"{}\", reference=\"{}\", table=\"{}\")".format(_inflector.classify(
                     table_relationship["table_name"]), table_relationship["column_name"], table_relationship["referenced_column_name"], table_relationship["table_name"]))
 
-                if table_relationship["table_name"] != table["TABLE_NAME"] and not table_relationship["table_name"] in referenced_table:
-                    entitie.write("\n\tdef {}(self): pass".format(
-                        table_relationship["table_name"]))
-                else:
-                    rel_name = table_relationship["constraint_name"].replace(table["TABLE_NAME"], "").replace("FK_", "").replace("fk_", "").replace("__","_")
-                    rel_name = rel_name[1:] if rel_name.startswith("_") else rel_name[:-1] if rel_name.endswith("_") else rel_name
-                    entitie.write("\n\tdef {}(self): pass".format(rel_name))
+                entitie.write("\n\tdef {}(self): pass".format(__table_referenced))
 
-                referenced_table.append(table_relationship["table_name"])
+                # if table_relationship["table_name"] != table["TABLE_NAME"] and not table_relationship["table_name"] in referenced_table:
+                    # entitie.write("\n\tdef {}(self): pass".format(
+                        # table_relationship["table_name"]))
+                # else:
+                    # rel_name = table_relationship["constraint_name"].replace(table["TABLE_NAME"], "").replace("FK_", "").replace("fk_", "").replace("__","_")
+                    # rel_name = rel_name[1:] if rel_name.startswith("_") else rel_name[:-1] if rel_name.endswith("_") else rel_name
+                    # entitie.write("\n\tdef {}(self): pass".format(rel_name))
 
             if len(table_relationships_n_to_n) > 0:
                 entitie.write("\n\n\t# Many-to-many")
 
             for table_relationship in table_relationships_n_to_n:
+
+                __table_referenced = reference_name(table["TABLE_NAME"], referenced_table, table_relationship["name"], table_relationship["constraint_name"])
+
                 entitie.write("\n\n\t@ObjectManyList(name=\"{}\", key=\"{}\", reference=\"{}\", table=\"{}\", intermediate=\"{}\", inter_key=\"{}\", end_key=\"{}\")".format(_inflector.classify(
                     table_relationship["name"]), table_relationship["key"], table_relationship["reference"], table_relationship["table"], table_relationship["intermediate"], table_relationship["inter_key"], table_relationship["end_key"]))
-                entitie.write("\n\tdef {}(self): pass".format(
-                    table_relationship["name"]))
+                entitie.write("\n\tdef {}(self): pass".format(__table_referenced))
 
             entitie.write("\n")
+
     # init model path
 
     # INICIA PASTA LIB ENTITY __INIT__.PY