+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Melhoria preenchimentos objetos relacionais
 * ORWHERE ou ORHAVING em Blocos
 * DataTypes Float e Decimal com round nas casas decimais
+* Delete Object e Delete Condition
 
 ### 0.0.4
 * Correção Comparação de Objetos e PK's Diferentes